feat: add Codex-compatible GET  and split handler into modules
