+++ crates/agentic-server/src/handler/http/mod.rs
@@ -1,0 +1,7 @@
+mod conversations;
+mod models;
+mod responses;
+
+pub use conversations::conversations;
+pub use models::{health, models, ready};
+pub use responses::responses;
