+++ crates/agentic-server/src/handler/mod.rs
@@ -1,0 +1,7 @@
+mod common;
+pub mod http;
+pub mod websocket;
+
+pub use common::{convert_response, executor_error_response};
+pub use http::{conversations, health, models, ready, responses};
+pub use websocket::responses_ws;
