+++ crates/agentic-server/src/handler/websocket/mod.rs
@@ -1,0 +1,4 @@
+mod error;
+mod responses;
+
+pub use responses::responses_ws;
